fixed some naming problems with utf16/ucs2

+++ src/host/Ucs2CharRow.hpp
@@ -31,7 +31,7 @@ Revision History:
 class Ucs2CharRow final : public CharRowBase<wchar_t, std::wstring>
 {
 public:
-    Ucs2CharRow(short rowWidth);
+    Ucs2CharRow(size_t rowWidth);
     Ucs2CharRow(const Ucs2CharRow& a) = default;
     Ucs2CharRow& operator=(const Ucs2CharRow& a);
     Ucs2CharRow(Ucs2CharRow&& a) noexcept = default;
